+++ crates/astria-sequencer/src/mempool/mod.rs
@@ -1,5 +1,5 @@
-// #[cfg(feature = "benchmark")]
-// mod benchmarks;
+#[cfg(feature = "benchmark")]
+mod benchmarks;
 mod mempool_state;
 mod transactions_container;
 
@@ -89,6 +89,7 @@ const REMOVAL_CACHE_SIZE: usize = 50_000;
 ///
 /// This is useful for when a transaction fails execution or when
 /// a transaction is invalidated due to mempool removal policies.
+#[cfg_attr(feature = "benchmark", derive(Clone))]
 struct RemovalCache {
     cache: HashMap<TransactionId, RemovalReason>,
     remove_queue: VecDeque<TransactionId>,
@@ -285,6 +286,13 @@ impl Mempool {
     pub(crate) async fn is_tracked(&self, tx_id: &TransactionId) -> bool {
         self.inner.read().await.is_tracked(tx_id)
     }
+
+    #[cfg(feature = "benchmark")]
+    pub(crate) async fn deep_clone(&self) -> Self {
+        Self {
+            inner: Arc::new(RwLock::new(self.inner.read().await.clone())),
+        }
+    }
 }
 
 #[derive(Debug)]
@@ -293,6 +301,7 @@ pub(crate) enum InsertionStatus {
     AddedToPending,
 }
 
+#[cfg_attr(feature = "benchmark", derive(Clone))]
 struct MempoolInner {
     pending: PendingTransactions,
     parked: ParkedTransactions<MAX_PARKED_TXS_PER_ACCOUNT>,